set1/challenge1: finish and pass challenge1
The issue before was that I assumed the string had to be read as a
string, which had characters representing hex values, which meant that
each character would have a 8-bit binary representation, since they are
all ASCII characters.

Instead it is passed as a string but is one big number in hex
representation, therefore each 'character' or digit, has a 4-bit binary
representation.

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>